backend/requirements.txt
amqp==5.2.0
asgiref==3.8.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.11
django-debug-toolbar==4.3.0
django-filter==24.1
djangorestframework==3.15.1
drf-yasg==1.21.7
flower==2.0.1
humanize==4.9.0
inflection==0.5.1
kombu==5.3.6
Markdown==3.6
packaging==24.0
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
tornado==6.4
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
